move topic grpc wrapper code to one file
