+++ tox.ini
@@ -46,12 +46,12 @@ deps =
 [testenv:black-format]
 skip_install = true
 commands =
-    black ydb examples tests --extend-exclude ydb/_grpc
+    black ydb examples tests --extend-exclude "ydb/_grpc/v3|ydb/_grpc/v4"
 
 [testenv:black]
 skip_install = true
 commands =
-    black --diff --check ydb examples tests --extend-exclude ydb/_grpc
+    black --diff --check ydb examples tests --extend-exclude "ydb/_grpc/v3|ydb/_grpc/v4"
 
 [testenv:pylint]
 deps = pylint
