+++ ydb/_topic_common/common.py
@@ -1,0 +1,3 @@
+import typing
+
+TokenGetterFuncType = typing.Optional[typing.Callable[[], str]]
